+++ wiki/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 Compilation history for sage-wiki.
 
+## 2026-06-10T22:06:56Z
+
+- Added: 29 sources
+- Modified: 0 sources
+- Removed: 0 sources
+- Summarized: 2
+- Concepts extracted: 1
+- Articles written: 1
+- Errors: 9
+
 ## 2026-06-10T19:51:25Z
 
 - Added: 114 sources